requirements.txt
# spatrack
easydict==1.13
opencv-python==4.9.0.80
moviepy==1.0.3
flow-vis==0.1
matplotlib==3.8.3
einops==0.7.0
timm==0.6.7
scikit-image==0.22.0
scikit-learn==1.4.1.post1
cupy-cuda11x
accelerate
yt-dlp
pandas

# cogvideox
bitsandbytes
diffusers>=0.31.2
transformers>=4.45.2
hf_transfer>=0.1.8
peft>=0.12.0
decord>=0.6.0
wandb
torchao>=0.5.0
sentencepiece>=0.2.0
imageio-ffmpeg>=0.5.1
numpy>=1.26.4
git+https://github.com/asomoza/image_gen_aux.git


# submodules
-r submodules/MoGe/requirements.txt